update theme service application
